+++ R/AllGenerics.R
@@ -159,3 +159,16 @@ setGeneric("ridgeplot",
            })
 
 
+#' upsetplot method generics
+##'
+##'
+##' @docType methods
+##' @name upsetplot
+##' @rdname upsetplot-methods
+##' @title upsetplot method
+##' @param x object
+##' @param ... additional parameters
+##' @return plot
+##' @export
+setGeneric("upsetplot", function(x, ...) standardGeneric("upsetplot"))
+